+++ tests/testthat/test-limit_skip_n.R
@@ -22,12 +22,12 @@ test_that("limit_n is robust", {
   expect_equivalent(vapour_read_names(f, skip_n = 56), 57:58)
   expect_equivalent(vapour_read_names(f, skip_n = 56, limit_n = 4), 57:58)
   expect_equivalent(vapour_read_names(f, skip_n = 56, limit_n = 1), 57)
-  expect_equivalent(vapour_read_names(f, sql = "SELECT FID FROM list_locality_postcode_meander_valley WHERE FID < 10"),
-                    1:9)
+  # expect_equivalent(vapour_read_names(f, sql = "SELECT FID FROM list_locality_postcode_meander_valley WHERE FID < 10"),
+  #                   1:9)
 
-  expect_equivalent(vapour_read_names(f, sql = "SELECT FID FROM list_locality_postcode_meander_valley WHERE FID < 10",
-                                      limit_n = 4, skip_n = 3),
-                    4:7)
+  # expect_equivalent(vapour_read_names(f, sql = "SELECT FID FROM list_locality_postcode_meander_valley WHERE FID < 10",
+  #                                     limit_n = 4, skip_n = 3),
+  #                   4:7)
   expect_equivalent(lengths(vapour_geom_summary(f,  limit_n = 4, skip_n = 9)), rep(4L, 7))
 
   expect_equal(lengths(vapour_geom_summary(f,  limit_n = 4, skip_n = 9)),